Mark all commands as internal and move its base to Commands namespace
